test: organiza execucao (base, models e flows web)
